+++ src/templates/html.php
@@ -8,8 +8,8 @@
 		<meta name="robots" content="%ROBOTS%" />
 		<meta http-equiv="content-type" content="text/html; charset=UTF-8" />
 		<meta http-equiv="X-UA-Compatible" content="IE=9; IE=8;" />
-		<meta name="author" content="Markdown | Notes management" />
-		<meta name="designer" content="Markdown | Notes management" />
+		<meta name="author" content="MarkNotes | Notes management" />
+		<meta name="designer" content="MarkNotes | Notes management" />
 		<meta name="keywords" content="%TITLE%" />
 		<meta name="description" content="%TITLE%" />
 		<meta name="apple-mobile-web-app-capable" content="yes" />