+++ src/templates/pdf.php
@@ -5,8 +5,8 @@
 
 		<meta charset="utf-8"/>
 		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-		<meta name="author" content="Markdown | Notes management" />
-		<meta name="designer" content="Markdown | Notes management" />
+		<meta name="author" content="MarkNotes | Notes management" />
+		<meta name="designer" content="MarkNotes | Notes management" />
 		<meta name="keywords" content="%TITLE%" />
 		<meta name="description" content="%TITLE%" />
 		<meta property="og:url" content="%URL_PAGE%" />