+++ src/templates/reveal.php
@@ -14,8 +14,8 @@
 		<meta name="robots" content="%ROBOTS%" />
 		<meta http-equiv="content-type" content="text/html; charset=UTF-8" />
 		<meta http-equiv="X-UA-Compatible" content="IE=9; IE=8;" />
-		<meta name="author" content="Markdown | Notes management" />
-		<meta name="designer" content="Markdown | Notes management" />
+		<meta name="author" content="MarkNotes | Notes management" />
+		<meta name="designer" content="MarkNotes | Notes management" />
 		<meta name="keywords" content="%TITLE%" />
 		<meta name="description" content="%TITLE%" />
 		<meta name="apple-mobile-web-app-capable" content="yes" />
@@ -32,8 +32,8 @@
         <title>%TITLE%</title>
 
         <link rel="stylesheet" type="text/css" href="%ROOT%/libs/reveal/css/reveal.css" media="screen" />
-		<link rel="stylesheet" type="text/css" href="%ROOT%/libs/reveal/css/theme/beige.css" id="theme" media="screen" />
-		<link rel="stylesheet" type="text/css" href="%ROOT%/libs/font-awesome/css/font-awesome.min.css" media="screen" />
+    		<link rel="stylesheet" type="text/css" href="%ROOT%/libs/reveal/css/theme/beige.css" id="theme" media="screen" />
+    		<link rel="stylesheet" type="text/css" href="%ROOT%/libs/font-awesome/css/font-awesome.min.css" media="screen" />
 
         <!-- Theme used for syntax highlighting of code -->
         <link rel="stylesheet" href="%ROOT%/libs/reveal/lib/css/zenburn.css">